Conversion projet test + MAJ scripts de build

+++ build.sh
@@ -1,4 +1,6 @@
-rm -r build
-rm -r docker/build
-dotnet publish src/todo -o ./build -f netcoreapp1.1
+[ -d "build" ] && rm -r build
+[ -d "docker/build" ] && rm -r docker/build
+[ -d "src/todo/build" ] && rm -r src/todo/build
+
+dotnet publish src/todo -o build -f netcoreapp1.1
 cp -a src/todo/build/. docker/build
